Return special characters. Fixes.
