+++ mne/datasets/spm_face/__init__.py
@@ -1,4 +1,4 @@
 """SPM face dataset
 """
 
-from .spm_data import data_path, has_spm_data, get_version
+from .spm_data import data_path, has_spm_data, get_version, requires_spm_data
