+++ mne/datasets/testing/__init__.py
@@ -1,5 +1,4 @@
 """MNE testing dataset
 """
 
-from ._testing import (data_path, requires_testing_data, get_version,
-                       requires_spm_data)
+from ._testing import data_path, requires_testing_data, get_version
